for #2567, move SelectStatement.getDistinctSelectItem to ShardingSelectOptimizedStatement
